wts2-1854/Baseline RNA-Seq Load

+++ bin/rnaseqBaselineDelete.sh
@@ -18,33 +18,8 @@ touch ${BASELINELOG}
 
 cat - <<EOSQL | ${PG_DBUTILS}/bin/doisql.csh $0 >> ${BASELINELOG} 2>&1
 
-select a.accid, m.* 
-into temp table baseline
-from MGI_Set s, MGI_SetMember m , ACC_Accession a
-where s.name = 'Baseline RNASeq Load Experiments'
-and s._set_key = m._set_key
-and s._mgitype_key = a._mgitype_key
-and m._object_key = a._object_key
-and a._logicaldb_key = 189
-and a.preferred = 1
-;
-
-create index idx on baseline (_object_key);
-
-select * from baseline;
-
 delete from GXD_HTSample_RNASeqCombined where _CreatedBy_key = 1673;
-
-delete from GXD_HTSample_RNASeqSet
-using baseline
-where baseline._object_key = GXD_HTSample_RNASeqSet._experiment_key
-;
-
-select b.*, s.*, m.*
-from baseline b, GXD_HTSample_RNASeqSet s, GXD_HTSample_RNASeqSetMember m
-where b._object_key = s._experiment_key
-and s._rnaseqset_key = m._rnaseqset_key
-;
+delete from GXD_HTSample_RNASeqSet where _CreatedBy_key = 1673;
 
 EOSQL
